extract redirect action
